centralized the distro actor logic
