+++ frontend/frontend/actors/Distro/DistroActor.tsx
@@ -1,40 +1,63 @@
 import React, { createContext, useContext } from 'react';
 import { Actor, HttpAgent } from '@dfinity/agent';
-import * as distro from '../../interfaces/distro'; // Import the distro interface and IDL
+import * as cycles from '../../interfaces/cmc/cmc';
+import * as ledger from '../../interfaces/ledger/index';
+import * as distro from '../../interfaces/distro/index';
 
-// Define the shape of our context
+const CANISTER_IDS = {
+  cycles: "rkp4c-7iaaa-aaaaa-aaaca-cai",
+  ledger: "ryjl3-tyaaa-aaaaa-aaaba-cai",
+  distro: "jeb4e-myaaa-aaaak-aflga-cai",
+};
+
 interface DistroActorContextType {
-  createDistroActor: (agent: HttpAgent) => Promise<any>;  // Function to create a distro actor
+  createActors: (currentUser: any) => Promise<any>;
+  createDistroActor: (agent: HttpAgent) => Promise<any>;
 }
 
-// Create a new context with the defined type
 const DistroActorContext = createContext<DistroActorContextType | undefined>(undefined);
 
-// Custom hook to use the DistroActor context
 export const useDistroActor = () => {
   const context = useContext(DistroActorContext);
-  // Ensure the hook is used within a DistroActorProvider
   if (!context) {
     throw new Error('useDistroActor must be used within a DistroActorProvider');
   }
   return context;
 };
 
-// DistroActorProvider component to wrap the app and provide the context
 export const DistroActorProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  // Function to create a distro-specific actor
+  const createActors = async (currentUser: { agent: any; }) => {
+    try {
+      const actors: { [key: string]: any } = {};
+
+      const canisters = [
+        { name: "cycles", id: CANISTER_IDS.cycles, factory: cycles.idlFactory },
+        { name: "ledger", id: CANISTER_IDS.ledger, factory: ledger.idlFactory },
+        { name: "distro", id: CANISTER_IDS.distro, factory: distro.idlFactory },
+      ];
+
+      for (const { name, id, factory } of canisters) {
+        actors[name] = Actor.createActor(factory, {
+          agent: currentUser.agent,
+          canisterId: id,
+        });
+      }
+
+      return actors;
+    } catch (error) {
+      console.error("Error creating actors:", error);
+      throw error;
+    }
+  };
   const createDistroActor = async (agent: HttpAgent) => {
-    // Create an actor using the distro IDL factory and a specific canister ID
-    const distroActor = Actor.createActor(distro.idlFactory, {
+    return Actor.createActor(distro.idlFactory, {
       agent,
-      canisterId: 'jeb4e-myaaa-aaaak-aflga-cai',  // Specific canister ID for the distro actor
+      canisterId: CANISTER_IDS.distro,
     });
-    return distroActor;
   };
 
-  // Provide the context value (createDistroActor function) to all children components
   return (
-    <DistroActorContext.Provider value={{ createDistroActor }}>
+    <DistroActorContext.Provider value={{ createActors, createDistroActor }}>
       {children}
     </DistroActorContext.Provider>
   );
